Add basic layout for "Archive" page
This page, as the summary says, should eventually contain descriptions
of the past meetups of the group. For now, I'm just planning on using
the Meetup.com page for reference, as I'm not aware of any other good
archive of this info, though the mailing list archives might contain
some details. The Meetup.com references go back to 2011, and thus so
does this page.

For now, at any rate, there's merely placeholders for the various years,
though. I'll add in the actual meetups shortly.

Also, I made the super-narrow version of the page (i.e., the one for
mobile browsers) shrink down the header text, as with the long
single-word title caused by "Baltimore::PM::Archive", it was running off
the screen and not looking very good.