Add more sidebar sections
There are now two new sidebar sections: Sponsors and "See Also". The
former is for groups which have a specific positive impact on our group,
which I plan to expand on in a new section shortly.

The other is, for now, just for our partner group DC.pm which helps us
put on DCBPW, as well as for DCBPW itself.