Add a "Sponsors" content section on the home page
This expands on the link from the side bar, explaining what it is that
each of these organizations does for us. I can sometimes veer into the
florid in terms of my prose when expressing gratitude, but hopefully I
have reined it in sufficiently so as not to be too much in this case.