Add 2011 archive of meetings
This is as far back as the Meetup.com history goes. There was a meeting
or two before this during @reyjrar's leadership period that I can find
on the mailing list, but not with sufficient detail to really fill
anything out (it might be possible to dig through this very repo to find
those details, the mailing list entries point to the website for the
agenda).

The mailing list was briefly active in 2005, but as far as I can tell,
there were no meetings held.

However, there were meetings back in 1999/2000 that were held during the
founding years of this group. These meetings occasionally list details
about speaker/topic, but it'd be somewhat difficult to parse, and I'm
unsure how many people will be interested beyond the history buffs.

That's not to say it wouldn't be worth doing, but for the moment I'm
just going to stick a fork in this and call it done.